Add logic package and logic component. Update DG
