+++ src/main/java/athena/logic/Logic.java
@@ -1,0 +1,7 @@
+package athena.logic;
+
+import athena.exceptions.CommandException;
+
+public interface Logic {
+    boolean execute(String inputString) throws CommandException;
+}
