+++ src/main/java/athena/logic/commands/DeleteCommand.java
@@ -1,4 +1,4 @@
-package athena.commands;
+package athena.logic.commands;
 
 import athena.exceptions.TaskNotFoundException;
 import athena.task.Task;
@@ -7,23 +7,23 @@ import athena.Ui;
 import java.util.Objects;
 
 /**
- * Handles the done command.
+ * Handles the delete command.
  */
-public class DoneCommand extends Command {
-    private int taskNumber;
+public class DeleteCommand extends Command {
+    private int deleteIndex;
 
     /**
-     * Initializes the object with the task number of task to be mark as done.
+     * Initializes the object with the index of task to be deleted.
      *
-     * @param taskNumber Integer representing the task number of task.
+     * @param index Integer representing the index of task.
      */
-    public DoneCommand(int taskNumber) {
-        this.taskNumber = taskNumber;
+    public DeleteCommand(int index) {
+        deleteIndex = index;
     }
 
     /**
-     * Marks a task as done from the Tasks list and
-     * calls Ui to print task marked as done.
+     * Deletes a task from the Tasks list and
+     * calls Ui to print task deleted.
      *
      * @param taskList Tasks list
      * @param ui       Ui
@@ -32,8 +32,9 @@ public class DoneCommand extends Command {
      */
     @Override
     public void execute(TaskList taskList, Ui ui) throws TaskNotFoundException {
-        Task taskDone = taskList.markTaskAsDone(taskNumber);
-        ui.printTaskDone(taskDone);
+        Task deletedTask = taskList.deleteTask(deleteIndex);
+        String taskRestore = deletedTask.getTaskRestore();
+        ui.printTaskDeleted(deletedTask, taskRestore);
     }
 
     /**
@@ -46,15 +47,15 @@ public class DoneCommand extends Command {
         if (this == o) {
             return true;
         }
-        if (!(o instanceof DoneCommand)) {
+        if (!(o instanceof DeleteCommand)) {
             return false;
         }
-        DoneCommand that = (DoneCommand) o;
-        return taskNumber == that.taskNumber;
+        DeleteCommand that = (DeleteCommand) o;
+        return deleteIndex == that.deleteIndex;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(taskNumber);
+        return Objects.hash(deleteIndex);
     }
 }
