+++ src/main/java/athena/logic/commands/ListCommand.java
@@ -1,4 +1,4 @@
-package athena.commands;
+package athena.logic.commands;
 
 import athena.Forecast;
 import athena.Importance;
