+++ src/main/java/athena/logic/commands/ViewCommand.java
@@ -1,29 +1,29 @@
-package athena.commands;
+package athena.logic.commands;
 
 import athena.exceptions.TaskNotFoundException;
-import athena.task.Task;
 import athena.TaskList;
 import athena.Ui;
+
 import java.util.Objects;
 
 /**
- * Handles the delete command.
+ * Handles the view command.
  */
-public class DeleteCommand extends Command {
-    private int deleteIndex;
+public class ViewCommand extends Command {
+    private int taskNumber;
 
     /**
-     * Initializes the object with the index of task to be deleted.
+     * Initializes the object with the task number of task to be viewed.
      *
-     * @param index Integer representing the index of task.
+     * @param taskNumber Integer representing the task number of task.
      */
-    public DeleteCommand(int index) {
-        deleteIndex = index;
+    public ViewCommand(int taskNumber) {
+        this.taskNumber = taskNumber;
     }
 
     /**
-     * Deletes a task from the Tasks list and
-     * calls Ui to print task deleted.
+     * View a task from the Tasks list and
+     * calls Ui to print task details.
      *
      * @param taskList Tasks list
      * @param ui       Ui
@@ -32,9 +32,8 @@ public class DeleteCommand extends Command {
      */
     @Override
     public void execute(TaskList taskList, Ui ui) throws TaskNotFoundException {
-        Task deletedTask = taskList.deleteTask(deleteIndex);
-        String taskRestore = deletedTask.getTaskRestore();
-        ui.printTaskDeleted(deletedTask, taskRestore);
+        String taskDescription = taskList.getTaskDescription(taskNumber);
+        ui.printTaskDetails(taskDescription);
     }
 
     /**
@@ -47,15 +46,15 @@ public class DeleteCommand extends Command {
         if (this == o) {
             return true;
         }
-        if (!(o instanceof DeleteCommand)) {
+        if (!(o instanceof ViewCommand)) {
             return false;
         }
-        DeleteCommand that = (DeleteCommand) o;
-        return deleteIndex == that.deleteIndex;
+        ViewCommand that = (ViewCommand) o;
+        return taskNumber == that.taskNumber;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(deleteIndex);
+        return Objects.hash(taskNumber);
     }
 }
