+++ src/test/java/athena/ParserTest.java
@@ -1,15 +1,16 @@
 package athena;
 
-import athena.commands.AddCommand;
-import athena.commands.Command;
-import athena.commands.DeleteCommand;
-import athena.commands.DoneCommand;
-import athena.commands.EditCommand;
-import athena.commands.ExitCommand;
-import athena.commands.HelpCommand;
-import athena.commands.ListCommand;
-import athena.commands.ViewCommand;
+import athena.logic.commands.AddCommand;
+import athena.logic.commands.Command;
+import athena.logic.commands.DeleteCommand;
+import athena.logic.commands.DoneCommand;
+import athena.logic.commands.EditCommand;
+import athena.logic.commands.ExitCommand;
+import athena.logic.commands.HelpCommand;
+import athena.logic.commands.ListCommand;
+import athena.logic.commands.ViewCommand;
 import athena.exceptions.CommandException;
+import athena.logic.Parser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
